refactor(core): database and logger module, app_config and swagger config moved to the folder core
