+++ src/app.module.ts
@@ -4,7 +4,7 @@ import { Module } from '@nestjs/common'
 import { ConfigModule } from '@nestjs/config'
 import { AppController } from './app.controller'
 import { AppService } from './app.service'
-import config from './config'
+import config from './core/app_config'
 import { OrderModule } from './modules/order/order.module'
 import { RestaurantModule } from './modules/restaurant/restaurant.module'
