+++ src/core/app_config.ts
