+++ src/modules/auth/auth.module.ts
@@ -1,9 +1,9 @@
-import { DatabaseModule } from '@modules/database/database.module'
-import { LoggerService } from '@modules/logger/logger.service'
 import { Module } from '@nestjs/common'
 import { ConfigModule, ConfigService } from '@nestjs/config'
 import { JwtModule } from '@nestjs/jwt'
 import { PassportModule } from '@nestjs/passport'
+import { DatabaseModule } from 'src/core/database/database.module'
+import { LoggerService } from 'src/core/logger/logger.service'
 import { AuthController } from './auth.controller'
 import { AuthService } from './auth.service'
 import { JwtAuthGuard } from './guards/jwt-auth.guard'
