+++ src/modules/client/client.service.ts
@@ -1,4 +1,3 @@
-import { LoggerService } from '@modules/logger/logger.service'
 import {
   BadRequestException,
   Inject,
@@ -7,6 +6,7 @@ import {
   NotFoundException
 } from '@nestjs/common'
 import { validate, ValidationError } from 'class-validator'
+import { LoggerService } from 'src/core/logger/logger.service'
 import { Repository } from 'typeorm'
 import { CreateClientDto } from './dto/create-client.dto'
 import { UpdateClientDto } from './dto/update-client.dto'
