+++ src/modules/order/order.module.ts
@@ -1,10 +1,10 @@
 import { ClientModule } from '@modules/client/client.module'
 import { ClientProvider } from '@modules/client/client.provider'
-import { DatabaseModule } from '@modules/database/database.module'
-import { LoggerService } from '@modules/logger/logger.service'
 import { RestaurantModule } from '@modules/restaurant/restaurant.module'
 import { RestaurantProvider } from '@modules/restaurant/restaurant.provider'
 import { Module } from '@nestjs/common'
+import { DatabaseModule } from 'src/core/database/database.module'
+import { LoggerService } from 'src/core/logger/logger.service'
 import { OrderController } from './order.controller'
 import { OrderProvider } from './order.provider'
 import { OrderService } from './order.service'
