+++ src/modules/restaurant/restaurant.module.ts
@@ -1,8 +1,8 @@
 import { ClientModule } from '@modules/client/client.module'
 import { ClientProvider } from '@modules/client/client.provider'
-import { DatabaseModule } from '@modules/database/database.module'
-import { LoggerService } from '@modules/logger/logger.service'
 import { Module } from '@nestjs/common'
+import { DatabaseModule } from 'src/core/database/database.module'
+import { LoggerService } from 'src/core/logger/logger.service'
 import { RestaurantController } from './restaurant.controller'
 import { RestaurantProvider } from './restaurant.provider'
 import { RestaurantService } from './restaurant.service'
